AIDEV/ai_system/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
torch>=1.9.0
transformers>=4.9.0
tensorflow>=2.6.0
matplotlib>=3.4.2
seaborn>=0.11.1
pytest>=6.2.5
pytest-cov>=2.12.1
pytest-asyncio>=0.15.1
black>=21.6b0
flake8>=3.9.2
mypy>=0.910
isort>=5.9.2
PyJWT>=2.1.0
cryptography>=3.4.7
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
prometheus-client>=0.11.0
psutil>=5.8.0
grafana-api>=1.0.3 